Fix %I macro expansion for IPv4 mapped IPv6 addresses.
On some systems (linux e.g.), when listening on ipv6 and ipv4,
addresses of ipv4 clients are printed as mapped ipv4 addresses
by getnameinfo (e.g. ::ffff:127.0.0.1). This re-establishes
the original behaviour of %I to expand to the plain ipv4 address
for an ipv4 client.

Michael
(This used to be commit 0abc8863f87fe5d2473492797c010784cb086008)