libreplace: fix samba4 build (by not setting global LIBS).
This corrects the earlier fix of the standalone build, by setting
LIBS to the desired value only in configure.ac but not in getifaddrs.m4.

Not that this changes the standalone build in that it adds these libs
undconditionally and not only if they are needed by the getifaddrs
replacement functions.

Michael
(This used to be commit ff3af1703ddc9e5383f32156bc5be8351f795e76)