libsmbconf: add internal open/close handling to registry backend.
This internally keeps track of opened registry in the private data
struct. The first call that really accesses data, opens the registry
and it is kept open until the destructor is called.

This behaviour might be changed in the future.

Michael
(This used to be commit 03e72e13076e3215eb8ae51cfb4e7cd3d3683d3e)