Fix libtdb some to move back towards allowing out of tree builds
Remaining problem is that not all *.syms linker scripts are generated
so the LD command needs to look in a single place.  Out of tree
builds now work but report write access to the source tree in order
to create the $(srcdir)/expotrs/libtdb.sysms file.
(This used to be commit 6e9287c5c3a6bbd15ec495593978c1a76e64ee60)