libsmbconf: add a text backend to libsmbconf based on the params.c parser.
This is a simple implementation of text backend for libsmbconf.
It is read only. It uses the parser from params.c. It loads the
given config file and holds the parsed lists of strings in memory.
Config is only re-read if the time stamp of the config file has changed.
Brilliant performance on large files is not to be expected since the
string lists are not hashed. This might be added in the future.

Michael
(This used to be commit 4ea50843852b335a9bd4c0535a5b684134ad8326)