Better doc mount.cifs.
Original commit msg from Jeff:

This patch adds a section to the mount.cifs manpage to clarify how
mount.cifs treats delimiters and when it autoconverts them from one
type to another. It also has it state that the use of forward slash
as a delimiter is preferred and why.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
(This used to be commit ec87ae251bea2a8b63febcf3ceab8754d564bdeb)