Enable multiple fd events on a single file descriptor
Now we can have multiple events listen on a single fd. The order in which they
are called is naturally not defined, meaning that it is only deterministic if
exactly one event is readable or writeable.
(This used to be commit 61503f822b60f98dd1f7549707ec89b48e32caeb)