Fixed a crash bug in unixuid module on failed ID mapping
We need to intialise *ids regardless of the status of the call, as the
IDL specifies that the out[] array always exists. If we don't
initialise out then we get a segv when a id mapping fails.

This still doesn't explain why the idmapping is failing, but at least
the client now gets NT_STATUS_NONE_MAPPED rather than a crashed
server.
(This used to be commit 4449ce381aca25e7f510a2f24b43c3a81e870032)