Far less cryptic traceback when you have an existing smb.conf
When the user has an existing smb.conf, but no [netlogon] or [sysvol]
share, the provision script would trigger a traceback.  While we still
need to abort in this situation, we do so now with a useful error.

Andrew Bartlett
(This used to be commit 10a8b7ea487f9725f69b02c4dd9cf5e1f67a23ab)