libsmbconf: add a smbconf init dispatcher smbconf_init().
The dispatcher takes a config source argument in the form
of "backend:path" where backend can (currently) be one of
"txt" (aka "file") or "reg" (aka "registry").

When the path is omitted, it is up to the backend to provide
a default path.

When there is no separator ":" and the string is not a recognized
backend, then file backend is assumed and the string is passed
in as the path argument.

Michael
(This used to be commit cc613a60eb0f6b92cae98e8669e5e569a1eb9324)