smbtorture: Add RAW-BENCH-LOOKUP micro-benchmark.
This test samples the lookup rate for a non-existant filename in a
directory, while varying the number of files in the directory. The
lookup rate should continue to approximate the lookup rate for the
empty directory case.
(This used to be commit ca7e4d9166a7cb10495cf3ecd9c4e72103fbec30)