When a test harness program fails, make the testsuite fail.
The problem fixed here is that pidl tests were not causing the 'number
of tests failing' count to increase, due to the way return codes are
processed on pipelines, in the shell.

By setting an exit code if we print 'failure', we ensure we fail
appropriately.

Andrew Bartlett
(This used to be commit 687e81883d37e3d1f55d3a7a87e20fb860888dde)