Add a new implementation of security=server.
This is not intended for general use, and will not be easily exposed
(if I have anything to do with it), but should allow the CIFS proxy to
re-use the connection.

A work in progress.

Andrew Bartlett
(This used to be commit a1e60ebc6d1e794011df5f69f691f4ec8622e991)