Don't block smbcontrol online/offline if !"winbind offline logons"
There's no point to block these, winbind will ignore them anyway.

Jerry, Jeremy, we definitely need a chance to trigger a reconnection attempt
after a DC has been rebooted. Right now winbind's reaction to a single rebooted
DC is to declare it offline. It will not try before a "winbind cache timeout"
timeout. If the admin sets this to an hour or longer for performance reasons,
we're screwed. Reboot your DC, no logon for an hour.

Volker