smbtorture: Add RAW-BENCH-TCON benchmark.
Add a simple test to benchmark the rate at which a server can accept
new tree connections. You can tune the length of time to run the
benchmark for and the number of parallel connections to make.
(This used to be commit ea3f4b93057e85c4ea516cc77dd0f293016d520c)