testsuite: add a testparm_s3 test script.
This is a first testparm/lp_load test that runs testparm on a couple
of configuration files. The main purpose for now is to test the options
that have special handlers (to check whether the handlers succeed).
In particular, all the Macro expansions that are available via
alloc_sub_basic() are tested with the include handler. This is to
catch such crashbugs as #5548 where %m expansion led to a segfault.

The tests now are very simple. Just check if testparm completes
successfully on the given config files. This can (and should) be
elaborated in the future.

Michael
(This used to be commit eb28146d40b8a8bc2c20b8d222abf191ea178d5a)