smbd: make sure the parent smbd stays alive forever
The problem was that the parent smbd doesn't have
any event when the process model standard was in use.

Now we always add an fd event for stdin, but may not
ask for any events to trigger.

metze
(This used to be commit fa6c00734ccf9bfe7a962253ddb5e2a1722c9b09)