Remove a check on p->vuid in _netr_LogonSamLogon/off
This is in preparation of removing p->vuid completely. I don't think this check
here makes any sort of sense, it only proves that the user having done the
session setup on its way to the schannel bind (ususally anonymous) did not do a
logoff in between. But this would have been caught way before this in
reply_pipe_write_and_X.