Remove events from require line:
The events library is not installed and prevent OpenChange from linking
with ldb. The following patch removes events reference from ldb pc file.

We are currently fixing this bug in openchange upstream using 'sed' in
our installsamba4.sh script, but this hack won't work with a packaged
version of samba4.

Signed-off-by: Stefan Metzmacher <metze@samba.org>