Collapse auxillary classes in LDAP schema conversion.
MS-ADTS 3.1.1.3.1.1.5 describes the behaviour of auxiliary classes.
In effect, these are additional MUST or MAY attributes that are
appeneded to the parent class (the auxiliary does not become listed in
the objectClass attribute), and so we do just that, and merge them
here, for export to OpenLDAP as it's schema.

Andrew Bartlett
(This used to be commit d674e92591ea90eb3b2117d8dd21f79f718d7730)