dcerpc.idl: the signing overhead can be more than 32 bytes
Windows seems to use 64 here, so we do now.

Before we got nca_proto_error fault because we send fragments
larger than the negotiated max frag size.

If the max frag size is 5840, we're sending 5837 bytes
when the auth_len is 45 and that matches w2k3 traffic.

metze
(This used to be commit 351947dba3f7a26ac871d4aa7b6bba4cd472383a)