libcli/raw: add a recv_helper hook infrastructure
The recv helper will be called when a response comes
and the recv helper can decide to let the request
on the SMBCLI_REQUEST_RECV when more reponse packets
are expected. It's up to the helper function
to keep a reference to the in buffers, each incoming
response overwrites req->in.

metze
(This used to be commit 6d84af89ba96627abe142ba7080c24ae2421ed6c)