build: set SONAME version of libraries from configure, not using library-versions
This sets the library versions (previously gotten from library-versions
file during make) already at configure time.

For library "foo", the version is stored in the LIBFOO_SOVER configure
variable.  For all libraries except libwbclient, this is done through the
call of SMB_LIBRARY().

NOTE: Currently the versions are hard-coded in the SMB_LIBRARY() calls in
configure.in, but this change makes it easier to retrieve the version from
the libraries source dir at configure time, e.g. for talloc and tdb.

Michael
(This used to be commit 3c300f9f8cffb5af32e291ec75007ab98fe96561)