tdbtool: fix off-by-one error in argument length. (bug #2344)
This prevented all commands operating on keys (all non-traverse commands)
in tdbtool to fail with a "fetch failed" or "delete failed" message.

It seems that it fixes bug #2344 ...

Apparently this bug was introduced with 94e53472666ed in 2005.
Either nobody is using tdbtool or else tdb_find() has become
more strict about the key legth in the meantime. :-)

Michael
(This used to be commit 9be6b46f7715cdc7f77308d7613e8df73effe3c1)