build: fix build of smbget by moving LIBSMBCLIENT_OBJ* defs up again
to their original place. The problem was that LIBSMCLIENT_OBJ1 is used
in SMBGET_OBJ and that the difinition of LIBSMBCLIENT_OBJ1 was moved
below the bin/smbget target. Somehow it had worked for me.

Thanks to Derrell for pointing that out.

Michael
(This used to be commit c4e51f5e761ab23fa08197fa9a52af6db4ab839d)