if a new node enters the cluster, that node will already be frozen at start but the rest of the nodes are not frozen.
at this stage an election is called by the new node.

Since in this case the nodes are not froze, we can not modify the recmaster
of the nodes so it is expected that this control would fail.

Add a boolean to send_election_request() to make it not
try to set the recmaster locally for the case where we are in an election phase
while not frozen.

(This used to be ctdb commit c5035657606283d2e35bea40992505e84ca8e7be)